+++ cypress/integration/search.spec.js
@@ -19,6 +19,15 @@ describe('Search', () => {
     });
   });
 
+  it('can search for flags bc clicking a category label', () => {
+    cy.visit('/flag/gay');
+    cy.wait(100);
+    cy.get('[data-cy-keyword="gay"]').last().click();
+    cy.location('search').should('eq', '?searchterm=gay');
+    cy.get("[data-cy-filter-options]").should('have.attr', 'data-cy-expanded', 'false');
+    cy.get('#search').should('have.value', 'gay');
+  });
+
   it('can find flags independent of case or spacing', () => {
     cy.get('#search').type('gay');
     cy.get('[data-cy-flagcard]').should('not.have.length', 0);