+++ src/data/stores/searchQuery.js
@@ -1,0 +1,5 @@
+import { writable } from 'svelte/store';
+
+export const searchterm = writable("");
+export const categoryFilters =  writable([]);
+export const colorFilters =  writable([]);