+++ src/routes/index.svelte
@@ -1,5 +1,10 @@
 <script>
+  import { onMount } from "svelte";
+
+  import { searchterm, colorFilters, categoryFilters } from './../data/stores/searchQuery';
+  
   import { app_name, app_baseUrl } from "../data/global.js";
+  import { filterCategories } from "./../data/_filter";
 
   import Metainfos from "../components/Helpers/Metainfos.svelte";
   import MainOptions from "../components/Filter/MainOptions.svelte";
@@ -20,17 +25,41 @@
     flag.keywords.push("valid");
   });
 
-  let activeColorFilters = [];
-  let activeCategoryFilters = [];
-  let searchterm = "";
+  let filterExpanded = false;
 
-  $: activeFilters = [activeColorFilters, activeCategoryFilters];
+  onMount(async => {
+
+    if ($categoryFilters.length || $colorFilters.length || $searchterm) {
+      filterExpanded = true;
+    }
+
+    const queryParams = new URLSearchParams(window.location.search);
+    if (queryParams.has('searchterm')) {
+      searchterm.set(queryParams.get('searchterm'));
+      categoryFilters.set([]);
+      colorFilters.set([]);
+    }
+    if (queryParams.has('category')) {
+      const paramCategory = queryParams.get('category');
+      const filterIds = filterCategories.map( (e) => { return e.id});
+      const filterExists = filterIds.includes(paramCategory);
+      if(filterExists) {
+        categoryFilters.set([queryParams.get('category')]);
+        colorFilters.set([]);
+        searchterm.set("");
+        filterExpanded = true;
+      }
+    }
+  });
+
+
+  $: activeFilters = [$colorFilters, $categoryFilters];
 
   let filtedFlags;
 
   $: searchresults_amount = filtedFlags.length;
 
-  $: if (activeFilters.flat().length || searchterm.length) {
+  $: if (activeFilters.flat().length || $searchterm.length) {
     filtedFlags = getFilteredFlags();
   } else {
     filtedFlags = flags;
@@ -38,19 +67,19 @@
 
   function getFilteredFlags() {
     let matchingFlags = flags;
-    if (activeColorFilters.length) {
+    if ($colorFilters.length) {
       matchingFlags = flags.filter(flag => {
-        return checkIfFlagMatchesColorFilters(flag, activeColorFilters);
+        return checkIfFlagMatchesColorFilters(flag, $colorFilters);
       });
     }
-    if (searchterm.length) {
+    if ($searchterm.length) {
       matchingFlags = matchingFlags.filter(flag => {
-        return checkIfFlagMatchesSearchRequest(flag, searchterm);
+        return checkIfFlagMatchesSearchRequest(flag, $searchterm);
       });
     }
-    if (activeCategoryFilters.length) {
+    if ($categoryFilters.length) {
       matchingFlags = matchingFlags.filter(flag => {
-        return checkIfFlagMatchesCategories(flag, activeCategoryFilters);
+        return checkIfFlagMatchesCategories(flag, $categoryFilters);
       });
     }
     return matchingFlags;
@@ -125,9 +154,7 @@
 <h1>Search</h1>
 
 <MainOptions
-  bind:activeColorFilters
-  bind:searchterm
-  bind:activeCategoryFilters
+  bind:expanded={filterExpanded}
   {searchresults_amount} />
 
 <div class="container">